HttpSM, HttpTunnel, and Http2Stream debug logging updates
When enabling debug logging with the http tag, it is often helpful to be
able to identify logs based on the corresponding state machine ID. This
is especially the case when there is a significant amount of concurrent
traffic such that many logs from different state machines are being
emitted at the same time. This patch updates logging calls to more
consistently add the associated state machine ID for such logs.