Traffic Dump: Use the correct transaction user index (#8548)
Traffic Dump has both session and transaction user data. In one of the
calls to TSUserArgGet, it accidentally used the session index to access
the transaction data. This could result in the corruption of another
plugin's session data. This patch fixes this so the correct index is
used.

(cherry picked from commit 42ad9468ba1bb755959f29230f1ae337f21caded)