Ports #7897 from core strategies to parent_select plugin.
Ports apache#7897 from core strategies to parent_select plugin.

Adds a new peering_ring mode to strategies, which allows parenting to
a consistent-hashed peer instead of the next hierarchical tier,
which can be useful in many situations, especially when the requesting
client itself is requesting different caches for the same key.