TSHttpTxnCacheLookupStatusGet: handle cannot respond cases (#8545)
ATS initially experiences cache lookups as a HIT for resources that we
ultimately cannot use due to a need for authentication or if the methods
for the incoming request and that of the request for the cached response
don't match. Our TSHttpTxnCacheLookupStatusGet plugin interface returned
the initial HIT status instead of marking these as MISSES. This patch
changes the status to return a MISS in these circumstances.

Closes #8539

(cherry picked from commit 080e236463490367ef94ce4e56fc95f287a917f5)