Update to Proxy Verifier version v2.3.0 (#8608)
This updates to the Proxy Verifier v2.3.0 release. This release updates
to the new libswoc::Errata and comes with some logging improvements.
Since the logs change, the test expectations for Proxy Verifier output
is updated with this commit as well.

(cherry picked from commit 9ca7618332673fdf2f86527e35fc304cf3989471)