is_stale_cache_response_returnable: verify HTTP method
This adds the method verification check currently in
is_cache_response_returnable to is_stale_cache_response_returnable. That
is, if an object is being requested by the client and the object is in
our cache, we can only reply with the cached object if the method of the
incoming request is the same as the method of the request for the cached
object.