Revert fixes for #8539
Fixing #8539 is easy to do in a common case, but it causing issues in
corner cases. For now I'm going to revert the fixes for this and circle
back around later to see whether this can be solved in a different way.

This reverts the following commits:

Revert "Commenting TSHttpTxnCacheLookupStatusGet need_to_revalidate (#8621)"
This reverts commit 4678ae34321002c584bd40ff6a1e4a128dae8bb1

Revert "TSHttpTxnCacheLookupStatusGet: call need_to_revalidate (#8617)"
This reverts commit 235c44a11f70d72168915af2d6504d0e63e880b3.

Revert "TSHttpTxnCacheLookupStatusGet: handle cannot respond cases (#8545)"
This reverts commit 080e236463490367ef94ce4e56fc95f287a917f5.