YAML: Define macro to hold the supported yaml version.
The idea is to remove this from the code and make it defined by the automake scripts.
Eventually the version could be read from the yaml source when added.
This is needed to validate the YAML version used by plugins when registering
to the JSONRPC manager to avoid ABI compatibility issues.