Revert remap.config loading changes
Revert "Short circuit remap reload when a valid remap file is not specified (#7782)"

This reverts commit d7847f23717d3a8be8341c244bbe1e27b2a6a77a.

This is being reverted from 9.2.x because changing remap.config loading
behavior was deemed too big a change to make within the 9.x releases.
See the discussion in #8802. These remap.config changes will be pursued
in the 10.x train.