Fixes issue #8807 nexthop failure threshold.
This was originally fixed in ATS 9.2.x with PR #8365 but #8365
was big and not backported to 9.1.x as it was a late PR and
zwoop did not want to bring it into 9.1.x at that time.

Anyway, this fixes the issue where the failure count on a parent
is not incremented properly.