remap_reload AuTest: await config reload messages
Change the remap_reload AuTest from waiting an arbitrary amount of time
(5 seconds) for the config reload to finish to instead waiting upon a
log message to indicate the reload has finished. This resolves some
flaky test results from the test when the previous 5 second sleep wasn't
enough time for the reload to finish.

Fixes: #8860