Adds a new member for TSRemapInterface tracking the plugin loaded
This allows for a plugin to introspect itself when TSRemapInit() is
being called. It's currently only implementing TSRemapDLHandleGet(),
which returns the handle to the .so as returned by dlopen(). With
this handle, a plugin can use e.g. dlsym() and dlinfo() to introspect
the plugin that it's loaded with.

Note that I've put this on 10-Dev. It's perhaps not ABI incompatible,
but seeing that it changes ts/remap.h internals, better to be safe
then sorry.