remap_reload test: Spacing out file modification and reload
The remap_reload.test.py AuTest is still flaky despite the last effort
to make it reliable. It looks like this might be because the
modification of the remap file and the reload is racing with each other.
This puts them in separate TestRuns and spaces them out by a few seconds
to see whether this makes the test more reliable.