Make clang-format not modify ink_autoconf.h.in and ink_autoconf.h (#8935)
ink_autoconf.h.in and ink_autoconf.h are generated files created with
autoconf. After they are made, if the user builds with them and then
calls the clang-format target, clang-format will dutifully format these
files according to our .clang-format specification. This then results in
a future invocation of make rebuilding most of the files in the
repository because those file are fundamental dependencies in our
repository.

This patch addresses the issue by simply not calling clang-format on
these generated, not checked in, files.

This fixes #8934