.fit/fmt/.clang-format-installed prerequisite
Before this patch, .git/fmt was used as the prerequisite to determine
whether clang-format was installed. But this directory was also created
by the autopep8 target, so if someone ran autopep8 before clang-format,
Make would think that clang-format was installed already. This adds a
hidden sentinel file to the clang-format install process that can be
used as a specific target in our clang-format Makefile.am.