Use sendmmsg and UDP GSO if available
If sendmmsg is unavailable, the code just calls sendmsg multiple times.

There is a case that UDP GSO is suppoted by Kernel but it's administratively
disabled or a network driver doesn't support it. The code detects such situation
on the first call for send(m)msg, and makes multiple msghdr on application side.