Update prepare_proxy_verifier.sh to work in non-git directories (#9048)
prepare_proxy_verifier.sh was written assuming that the user was in a
git repository. The user, however, may have a source code only copy of
the tree and thus may not be in a git repository. This makes the script
versatile to handle both of these environments.

Fixes #9045

(cherry picked from commit a40bc0f6658cdb059ec3169bf305dd419020a626)