Disable the QUIC TestRun in active_timeout AuTest (#9069)
We have added QUIC version 0x1 support to our 10-Dev branch via #8956,
but that has not been backported to 9.x and we don't plan to backport
that at this time. ASF CI will be updated with the latest version of
curl which uses 0x1 and would therefore fail. This change disables the
QUIC TestRun of active_timeout so it doesn't run on 9.x since ATS
doesn't support the latest QUIC test tools on that branch. This patch,
therefore, should not go into 10-Dev.

Note that this patch can be reverted in the future if we decide to
backport QUIC 0x1 support to 9.x.