Fix HTTP/2 session receive window handling for small sizes (#9117)
This fixes the way we handle session window sizes if the user configures
proxy.config.http2.initial_window_size_in to be smaller than the HTTP/2
protocol's default value of 65,535. For streams, we can communicate to
the client a smaller stream size via a SETTINGS frame, but not so for
smaller session window sizes. The way HTTP/2 hosts are expected to
reduce their receive session window size is by receiving DATA frames
without replying with WINDOW_UPDATE frames until the window falls to the
desired value. ATS already handled this latter behavior correctly
(letting the window size fall without sending WINDOW_UPDATE frames for
the session until it reached the desired smaller value). However, it
incorrectly reported an error and sent a GOAWAY frame when the user sent
DATA content above the reduced window size before ATS gave the client a
chance to reduce it via DATA frames. This fixes the receive session
window check to allow for a shrinking session window.

Fixes #9115