Reply with a 503 response when per_server.connection.max is exceeded
When the number of origin connections to a server exceeds
proxy.config.http.per_server.connection.max, we should reply with a 503.
This is done on master, but after a refactoring in 10-Dev, this was
missed. This adds the call back in that replies with a 503 when the max
is reached.

Fixes #9120