AuTest automatic keylog file configuration
Some tests have ad hoc TLS keylogging configured in case it is helpful
for analyzing TLS traffic later. This modifies the AuTest Traffic Server
process extension to automatically configure keylogging and to place it
in the Traffic Server process's log directory, which should be available
to the process in all CI environments.