Traffic Dump: fix YAML format for CONNECT requests (#9139)
For CONNECT requests, Traffic Dump will not have a server-side protocol
stack because we populate that on HTTP responses from the server. Since
we simply establish a connection to the origin without exchanging an
HTTP message, the stack is never populated. This resulted in malformed
YAML because we started a sequence with a ','. This patch addresses this
by not adding the ',' if the protocol description string is empty.

(cherry picked from commit 96acfe8192213c0af6a1df81c0c70417b505a015)