AuTest: Update to Proxy Verifier v2.5.0 (#9221)
v2.5.0 includes support for the await directive in which HTTP/2 streams
that depend upon the completion of a previous stream can list those
streams as a dependency. It also includes the ability to specify
RST_STREAM frames.

The ability to await a stream is needed for the HTTP/2 to origin PR
(#8963).