Docs: Unpin Sphinx
When Sphinx 4.x was initially released there were rendering issues with
it building the ATS docs. To address this we simply pinned to the 3.x
releases. Currently the latest Sphinx release is v6.1.3 and I noticed
that unpinning doesn't have issues with our code anymore. This therefore
unpins Sphinx. This also addresses a deprecated use of set_class from
traffic-server.py.