Update h2disable tests for new curl output
The h2disable and h2disable_no_accept_threads autests both had curl output
expectations that have changed in the recent release of curl. Here is
the old output:

```
Using HTTP2, server supports multiplexing
```

Here is the new output

```
using HTTP/2
```

This updates the ContainsExpression and ExcludesExpression strings to be
able to work with both versions of curl.