Proxy Protocol: Allow for no Proxy Protocol header (#9383)
Prior to this patch, if the user specified that a listening port may
have Proxy Protocol headers on connections via `pp` in the server_ports
configuration, the connection would be dropped if such a header was not
sent. This makes ATS flexible in such situations: if the Proxy Protocol
port is there, then it is processed, otherwise the connection is
processed according to the other specifications in the server_ports
configuration for the port.