http2_flow_control.test.py: Correctly enable debug logs
The http2_flow_control.test.py AuTest was ported from 10.x which
supports a value of 3 for `proxy.config.diags.debug.enabled`. 9.2.x,
however, does not support 3 and therefore debug logging hasn't been
working for this test in 9.2.x. This fixes that. This, incidentally,
seems to make the test run more reliably in the opensource CI
environment.