Remove premature context updates in TSSslSecretSet().
This PR partially reverts PR #8368.  SSL contexts (SSL_CTX) should only be updated by TSSslSecretUpdate()
(which is only called for the main, not the related secret).  The update in TSSslSecretSet() is not merely redundant,
it causes errors.  When an update is done, OpenSSL will check that the cert public key matches the private key.  Since
TSSslSecretSet() can only set one at time, if it also updates, they will not match.

Contexts for a loading confguration never have to be updated, becasue no SSL_CTX's have been created using it yet.