Remove window check from update_write_request
If we short circuit our update_write_request based upon HTTP/2 windows,
then we can get into a situation where we are not flushing DATA frames
to the client, but are internally accounting for them. Eventually we
stop trying to flush because our window is depleted, which results in us
never flushing. This manifested itself as DATA frames getting stuck in
our VIO buffers.