Fix clang 16 warnings (#9586)
* Address QUICPacket overloaded-virtual warnings

clang-16 warns that a couple functions in QuickPacket.h have the same
name as a virtual function but are not overloading. It is highlighting
that this might be a mistake. It is not in these cases, so we are
addressing this with using directives to make the intention explicit.

* Disable type-limits warning in Magic++.h

Magic++.h is a system file that has type warning in it. Disabling this
warning for that file because it is a system file rather than an ATS
file that we manage.

* Update query_remap.c return type

Fix int/enum type mismatch in query_remap.c by using the TSReturnCode
types.

* WSBuffer.h: Addressed undefined uint warnings

Include cstdint to include the definition of uint16_t.

* TSVConnFd.cc: include cstdint for int64_t definition