Ensure a reason phrase when sending an HTTP/1 response (#9615)
With an HTTP/1 client and an HTTP/2 origin, responses will come from the
origin without reason phrases because reason phrases are explicitly
removed from the HTTP/2 RFC. When expanding test coverage of reason
phrases it was noticed that when converting responses from HTTP/2 to
HTTP/1 to send toward the client, the reason phrases were empty. This
updates the conversion logic to ensure that HTTP/1 clients receive a
reason phrase.