Fix Proxy Protocol outbound
The Proxy Protocol outbound functionality broke with the HTTP/2 to
origin merge (#9366). I didn't update the Proxy Protocol feature
correctly with the merge and the CI's proxy_protocol.test.py was skipped
because our CI system didn't have the new go-httpbin in the path. This
fixes the outbound Proxy Protocol feature again.