rename attempts to retry_attempts (#9655)
From the name, the HttpTransact::_CurrentInfo.attempts seems to represent the total connection attempt, but it's currently really the retry attempt:

* it's initialized to 0
* it's incremented only in the OS or parent retry code path

This PR renames this to HttpTransact::_CurrentInfo.retry_attempts to avoid confusion. Note that there is no functionality changes. Changes are to just the naming and some related log messages(plus one assertion).