Add a new configuration to ignore query params in cache keys
This adds a new configuration option:

    CONFIG proxy.config.http.cache.ignore_query INT 0

This is off by default, but it is overridable. The typical use case
for this would be to use e.g. conf_remap.so plugin, to set this for
services / URLs which shouldn't use the query parameters in the HTTP
cache key calculations.

This would be significantly less expensive (less CPU) than using e.g.
the cachekey.so plugin to simply strip the query parameters from the key.