Disable proxies when running autest
When HTTP proxies are enabled on the system, autest would fail to run properly
because requests should never leave localhost. Now, proxies can be disabled by
unsetting the relevant environment variables in the CLI. However, it's not ideal
in this case because the autest script uses pipenv, which fetches Python
packages online, to run the actual tests.

This change fixes that by unsetting the environment variables after pipenv is
finished.