Break the circular dependency between iocore and tscore.
The re-introduction of the MTHashTable class created a dependency from
tscore to iocore, since it uses proxy mutexes, which are tightly coupled
to the event system. We can break this circular dependency by moving
the class into the QUIC library as an implementation detail, since that
already depends on the event system

This fixes #9768.

Signed-off-by: James Peach <jpeach@apache.org>