Add a shutdown handler to the certifier plugin
This adds a shutdown handler to the certifier plugin so that its SSL
data can be cleaned up in a callback rather than on process shutdown.

This also reverts commenting out the debug call from SslData which was
commented out as a workaround to a crash which was happening on
shutdown.

Fixes #9794