Move Catch from tests/include to lib/catch2 (#9815)
* Move catch.hpp from tests/include to lib/catch2

tests/ contains autests, not unit tests, so the catch header
shouldn't be isolated there. It is easier to locate in lib/
and will allow all potential CMake targets for bundled third-party
software to be in one place.

* Add copy of Catch2 license

Albeit not required, it is handy to have a copy of the license for
easy reference. Otherwise we have to open the header file and go to
a URL to find out which license it uses.