Resolve circular dep in esi components (#9817)
Some files in esi/fetcher depended on files in esi/lib and vice-versa.
I was able to move the shared dependencies into esi/common, and now
the dependency hierarchy is a little cleaner. This made it easy to
add the esi plugin and its tests to the CMake build, so this refactoring
is a precursor to bringing in those changes as well.