Add header_rewrite plugin to CMake build (#9890)
* Add header_rewrite plugin to CMake build

This builds the header_rewrite plugin and registers its associated
test with CTest. It looks for maxmindb's config file, which they
export since libmaxmindb-1.5.0, and links their target if found,
but it doesn't seem to be provided by OpenSUSE's libmaxmindb-devel-1.7.1
package.

* Add Find module for maxminddb

This ensures that we can build the header_rewrite plugin with
maxminddb on systems that don't correctly package the official
export.