Add support for quic to CMake build (#9892)
This adds an option that defaults to OFF: ENABLE_QUICHE. When quiche
is enabled, the quic static library and the http3 static library
will be built and linked appropriately. The traffic_quic executable
will also be built.

It also does some cleanup of targets in the CMake build. Some
targets are now namespaced in ts:: to make it explicit which targets
are part of trafficserver, and to make usage style consistent with
that of third-party libraries. Some dependencies have also been
corrected, and comments added where dependencies are misrepresented
by necessity.