Export traffic_server symbols (#9903)
This is a temporary fix until we can separate the API from the
traffic_server executable. It sets the ENABLE_EXPORTS property
on the traffic_server target to export its symbols, and "links"
all the plugins to it. On Linux, this does nothing of interest.
On Mac, the -bundle_loader flag will be added.

Closes #9848