Fix autests for fedora 38 (#9881)
I ran autest on fedora:38 and noticed that all but three passed. This commit updates our autests to work with the newer openssl s_client and Python3.11.

The most significant set of changes come from hyper being abandoned an no longer working on Python 3.11. This converts the ad-hoc clients we had written in hyper to the h2 framework.