OpenSSL 3.0 tls autest updates (#9947)
This updates the curl invocations for the tls_client_versions and
tls_client_versions_minmax autests so that it will run with the updated
OpenSSL 3.0 which is more strict. This avoids the issue by toning down
curl's security level to 0. A future PR should assess our
proxy.config.ssl.server.cipher_suite configurations and update
accordingly.

This also fixes a crash that tls.test.py found for OpenSSL 3.0 in which
SSLNetVConnection::load_buffer_and_write should return a non-zero in the
event of SSL_ERROR_SYSCALL.