Add interface for NetVC services (#9482)
It turned out that the use of dynamic_cast unignorably affects performance. #8003

A couple of approaches were suggested where we need to remove dynamic_cast for better performance and also want to avoid bringing much dependency into NetVConnection, and we took an approach that stores pre-casted pointers for mixins into a simple array. This avoids the use of RTTI when it access the mixins, and also allows us to use composition pattern, which can reduce the sizes of base classes.

Since the dynamic_casts were added separately little by little, we don't know how much it decreased the performance, but this change increased RPS by 4%.

Co-authored-by: James Peach <jpeach@apache.org>