Extract HttpUserAgent class from HttpSM (#9869)
* Extract HttpUserAgent class from HttpSM

This extracts a new class with 3 fields (entry, raw_buffer_reader,
and txn). It doesn't have much responsibility yet. But it does
provide a starting point for moving some behavior out of the
state machine.

* Move virtual connection stuff into HttpVCTable.h

This introduces HttpVCTable.h and HttpVCTable.cc to hold all the
stuff related to the HttpVCTable struct.

* Move a bunch of client_ fields into HttpUserAgent

This moves all the client_ fields except the byte counts into
HttpUserAgent, and sets most of the fields when the transaction
is set.

* Reset some client fields to defaults in set_txn

This matches the behavior of the original code, and may be
important.

* WIP Fix incorrect tcp_reused detection

A not (!) operator was missed, which caused tcp_reused to be
set incorrectly on the user agent. This fixes the issue, and adds
Catch tests to check for regression.

* Clean up debugging changes

There is one segfault left to fix.

* Fix format

* Fix proxy/http Makefile

* Add test_error_page_selection to CMake

* Fix rebase errors

* Add a getter for the state machine's user agent

* Test against UserAgent instead of HttpSM

* Move test_HttpSM to test_HttpUserAgent

* Do not build test_HttpUserAgent in automake build

* Add new sources back to automake