Shutdown tests: update backing sleep command to 10
The Ready conditions for some await processes waited upon the existence
of a log entry before continuing. This works fine, but the process
behind some of them was a 1 second sleep. If the sleep finished before
the file existed, which should be rare for these tests but possible, the
test would fail because the process finished before the process was
"ready". This updates the backing sleep commands to 10 second sleeps
which should afford plenty of time for the watch processes to see their
file content. This may make these tests more reliable in CI.