autest get_port: update to use psutil
This updates the get_port logic to pre-filter possible ports via psutil.
This should work quite effectively on Linux systems which can gather all
listening ports on the system via that module. On Mac, psutil is limited
to the user's processes, but it should still help there. Note that the
previous logic is kept, it's just that whatever ports are listening via
psutil are not considered available before using that logic.

I'm hopeful this will make some of the autests in CI more reliable.