polite_hook_wait logging updates
The polite_hook_wait.test.py autest frequently fails in the CI jenkins
jobs due to a hang loading the polite_hook_wait plugin. The CI debug
logs are empty in these situations. This patch adds more logs and
switches the logging mechanism to TSDebug in the hopes that this
provides us more information. We can revert this patch or parts of this
patch later once we figure out what is going wrong.