Pull INKAPIInternal out of InkAPI.cc as inkcont (#10114)
This is the first of three parts of a refactoring to break out FeatureAPIHooks,
making it possible to unit test hook callbacks.

This moves the implementation of INKAPIInternal from InkAPI.cc to
InkAPIInternal.cc, still in the traffic_server directory. For CMake,
I wrapped it in a new library, inkcont, as a convenient way to document
its dependencies and link it into unit tests. For autotools, I have
just added the new source file to the binary