Eliminate function-scope thread_local in Regex.cc.
It appears to be the case that, in the C/C++ runtime for Linux,
there is a mutex that is locked both:
1)  When initializing non-local variables in a shared library,
while it is being loaded by calling dlopen(), and
2)  When execution encounters the definition of a thread_local
variable within a function.  It seems that the mutex is locked
even if the variable has already been initialized, presumably to
check whether it was initialized.

This has been causing problems for the DbgCtl class.  When an
instance of DbgCtl is initialized, it locks a mutex (for the
registry of instances).  The instance initialization then
(indirectly) calls a function (in Regex.cc) that defines a
thread_local varaiable.  The result is scenarios that sometimes
lead to deadlock.  For example:
1)  dlopen() is called to load a plugin.  The runtime's mutex
is taken.
2)  Another thread starts initializing a DbgCtl instance, and
takes the instance registry mutex.
3)  The non-local initialization of the shared library starts
the intialization of another DbgCtl instance.  This thread now
blocks trying to lock the registry mutex.
4)  The other thread now enters the function in Regex.cc that
defines a thread_local object.  It blocks on the runtime's
mutex, which it needs to lock in order to do/check the
thread_local instances.  Deadlock.

This PR fixes the problem, by eliminating the thread_local
variable that's defined in a function.