Add AutoOptionHelpers.cmake
This module assists with making the usage of optional libraries
consistent and in agreement with automake. It provides functions
to register new options, and to print out the state of all associated
features.

Each option based on a package will follow the format ENABLE_XXX,
and will have an associated feature variable (e.g. TS_HAS_JEMALLOC).
Its value can be ON, 1, TRUE, OFF, 0, FALSE, or AUTO.

When the states of the features are printed out, features that are
ON or OFF will generally be color coded green or red respectively to
help configuration mistakes stand out. Options that were explicitly
turned off are given special treatment: if the feature is off, it
will not be color coded, because this is expected; if the feature is
found to be enabled despite the user's request, it will be colored red
and a CMake warning will be issued. All of this should help keep things
consistent and avoid any potentially annoying bugs caused by misconfiguration.